feat: comprehensive security hardening and HTTP/2 flow control fixes
Security Hardening:
- Fix CIDR whitelist bypass in SocketSYNProtect (invalid prefix validation)
- Add HTTP/2 flow control overflow protection (RFC 9113 §5.2.1)
- Implement Rapid Reset mitigation (CVE-2023-44487) with RST_STREAM rate limiting
- Add HPACK bomb prevention with expansion ratio checks
- Improve hash collision resistance with randomized seed in SYN protection
- Arena memory safety: disable LRU eviction in arena mode

New Features:
- Add structured logging context with SocketLogContext and SocketLogField
- Add new HTTP/2 metrics (FLOW_OVERFLOW, FLOW_NEGATIVE, HPACK_BOMB_DETECTED)
- Add SocketPool stats API for monitoring

Bug Fixes:
- Fix metrics precision in SocketSYNProtect_stats
- Fix duplicate gauge decrement in cleanup
- Optimize whitelist_check to parse IP only once per check

Testing & Documentation:
- Add 20+ new test cases for edge errors and flood scenarios
- Add benchmark_synprotect.c for performance measurement
- New fuzz harnesses: fuzz_dns_inj, fuzz_hpack, fuzz_http_smuggling, etc.
- New docs/SYN-PROTECT.md with usage and config tuning
- Update HTTP.md, SECURITY.md with mitigation best practices
- Regenerate Doxygen documentation

+++ include/core/SocketTimer-private.h
@@ -20,7 +20,7 @@ struct SocketTimer_T
     SocketTimerCallback callback; /* User callback function */
     void *userdata;             /* User data for callback */
     int cancelled;              /* Lazy deletion flag */
-    unsigned id;                /* Unique timer ID */
+    uint64_t id;               /* Unique timer ID (64-bit to prevent wrap-around) */
 
     size_t heap_index;          /* Heap position for fast lookup/cancel, SOCKET_TIMER_INVALID_HEAP_INDEX if not in heap */
 };
@@ -31,7 +31,7 @@ struct SocketTimer_heap_T
     struct SocketTimer_T **timers;  /* Dynamic array of timer pointers (heap) */
     size_t count;                   /* Current number of timers in heap */
     size_t capacity;                /* Allocated capacity of timers array */
-    unsigned next_id;               /* Next timer ID to assign */
+    uint64_t next_id;              /* Next timer ID to assign (64-bit to prevent wrap-around) */
     Arena_T arena;                  /* Memory arena for allocations */
     pthread_mutex_t mutex;          /* Thread safety mutex */
 };
